% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChatRbot.R
\name{api_call_chatGPT}
\alias{api_call_chatGPT}
\title{Send an API call to chatGPT}
\usage{
api_call_chatGPT(
  .messages,
  .openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  .model = "gpt-3.5-turbo",
  .max_tokens = 512L,
  .temp = 0.7,
  .top_p = 1,
  .presence_penalty = 0,
  .frequency_penalty = 0,
  .max_retries = 3,
  .v = FALSE,
  .timeout = 25
)
}
\arguments{
\item{.messages}{An message list in a predefined format (list with role and content information for each meassage)}

\item{.openai_api_key}{The API Key for OpenAI. By default loads one from Sys.getenv("OPENAI_API_KEY") - Set it in your .RProfile}

\item{.model}{The ChatGPT  modelto be used for ChatRBot. The default is "gpt-3.5-turbo",}

\item{.max_tokens}{The maximum number of tokens to generate in the completion.}

\item{.temp}{What sampling temperature to use, between 0 and 2. Higher values make the output more random. With 0 the output is fully deterministic, but likely less good.}

\item{.top_p}{The alernative sampling approach to temperature - set a return probability  between 0 and 1}

\item{.presence_penalty}{Between -2.0 and 2.0. Positive values -> new tokens penalized if they appear in the text so far}

\item{.frequency_penalty}{Between -2.0 and 2.0. Positive values -> new tokens penalized by frequency in the text so far}

\item{.max_retries}{#Retries after failed requests}

\item{.v}{#Do we want detailed connection details and metadata for each request}

\item{.timeout}{#Seconds after a request times out (Set higher for long requests)}
}
\value{
A message list in the same format as .messages
}
\description{
This function sends an API call to chatGPT and retries to send out the call if it did not work
}
