% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChatRbot.R
\name{mc_answer}
\alias{mc_answer}
\title{Get multiple choice answers from Claude models}
\usage{
mc_answer(
  .prompt,
  .choices,
  .imagefile = NULL,
  .capture_plot = FALSE,
  .model = "claude-3-opus-20240229",
  .temperature = 0
)
}
\arguments{
\item{.prompt}{A character string representing the question or prompt.}

\item{.choices}{A character vector of choices for the given prompt.}

\item{.imagefile}{An image to include into the user prompt}

\item{.capture_plot}{Should the last plot be captured and sent to Claude? (Default: FALSE)}

\item{.model}{A character string specifying the Claude model to use (default: "claude-3-opus-20240229").}

\item{.temperature}{A numeric value controlling the randomness of the generated answer (default: 0.0).}
}
\value{
A character string representing the best answer chosen by the Claude model.
}
\description{
This function sends a prompt with multiple choices to the Claude API and retrieves the best answer.
}
\examples{
mc_answer("What is the capital of France?", c("London", "Paris", "Berlin", "Madrid"))
}
