% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChatRbot.R
\name{chat}
\alias{chat}
\title{Chat with one of the Claude models}
\usage{
chat(
  .prompt,
  .imagefile = NULL,
  .pdffile = NULL,
  .capture_plot = FALSE,
  .f = function() {
     cat("")
 },
  .model = NULL,
  .max_tokens = 1024,
  .v = FALSE,
  .timeout = 60,
  .system = NULL,
  .temperature = 0.7,
  .c = NULL,
  .print = TRUE,
  .auto_save = TRUE,
  .return_history = FALSE
)
}
\arguments{
\item{.prompt}{The users current}

\item{.imagefile}{An image to include into the user prompt}

\item{.pdffile}{Can add the text of a PDF to the prompt}

\item{.capture_plot}{Should the last plot be captured and sent to Claude? (Default: FALSE)}

\item{.f}{Optional place to add a function, that gives it console output directly to Claude}

\item{.model}{The claude model to be used for ChatRBot. The default is "claude-3-opus-20240229",}

\item{.v}{Make the current request verbose}

\item{.timeout}{How long to wait after the request times out}

\item{.c}{Specify a conversation that is used to save messages or that is continued if it is pre-existing}

\item{.print}{Should the conversation be printed (Default: TRUE)}

\item{.sys}{The system message. Basic setup of Claude's role.}

\item{.auto_save_c}{Should the conversation be auto-saved (only applies if .c is not set manually)}

\item{.temp}{What sampling temperature to use, between 0 and 2. Higher values make the output more random. With 0 the output is fully deterministic, but likely less good.}
}
\description{
Chat with one of the Claude models
}
